qM3C2 on windows plugin mode
