+++ cmake/CloudCompareVariables.cmake
@@ -2,6 +2,10 @@ if(WIN32)
   set(CLOUDCOMPARE_EXECUTABLE_INSTALL_PATH
       ${CMAKE_INSTALL_PREFIX}/${CLOUDCOMPARE_DEST_FOLDER}/CloudCompare.exe
   )
+
+  set(CC_PYTHON_ENV_NAME "Python")
+  set(CC_PLUGIN_INSTALL_DIR ${CLOUDCOMPARE_DEST_FOLDER}/plugins)
+  set(CC_PYTHON_INSTALL_DIR "${CC_PLUGIN_INSTALL_DIR}/${CC_PYTHON_ENV_NAME}")
 elseif(APPLE)
   set(CLOUDCOMPARE_EXECUTABLE_INSTALL_PATH
       ${CMAKE_INSTALL_PREFIX}/CloudCompare/CloudCompare.app
