+++ cmake/PythonEnvHelper.cmake
@@ -23,7 +23,7 @@ macro(set_python_executable_from_current_venv)
       list(INSERT CMAKE_PREFIX_PATH 0 "$ENV{CONDA_PREFIX}/Library/share/cmake")
       set(PYTHON_EXECUTABLE "$ENV{CONDA_PREFIX}/python.exe")
     elseif(DEFINED ENV{VIRTUAL_ENV})
-      set(PYTHON_EXECUTABLE "$ENV{VENV_PREFIX}/Scripts/python.exe")
+      set(PYTHON_EXECUTABLE "$ENV{VIRTUAL_ENV}/Scripts/python.exe")
     endif()
   else()
     if(DEFINED ENV{CONDA_PREFIX})
