+++ wrapper/pycc/src/pycc.cpp
@@ -17,9 +17,6 @@
 
 #include <CCTypes.h>
 
-#include <QCoreApplication>
-#include <QDir>
-#include <QDirIterator>
 #include <QMainWindow>
 #include <QOpenGLWidget>
 #include <QProgressDialog>
@@ -43,6 +40,7 @@ using namespace pybind11::literals;
 
 void define_qcc_db(py::module &);
 void define_qcc_io(py::module &);
+void load_pluginWrappers(py::module &);
 
 template <class T> using observer_ptr = std::unique_ptr<T, py::nodelete>;
 
@@ -59,16 +57,6 @@ typename std::enable_if<std::is_same<T, float>::value, void>::type define_scalar
     m.attr("ScalarType") = py::dtype("float32");
 }
 
-QString guessPlatlibdir() {
-   #if PY_MAJOR_VERSION >= 3 && PY_MINOR_VERSION >= 9
-        py::module_ sys = py::module_::import("sys");
-        const QString platlibdir = sys.attr("platlibdir").cast<QString>();
-    #else
-        const QString platlibdir = "lib";
-    #endif
-    return platlibdir;
-}
-
 // Some plugins in CloudCompare are written to always display some
 // kind of dialogs.
 //
@@ -82,7 +70,15 @@ QString guessPlatlibdir() {
 void InitializeQApplication()
 {
 #ifdef PYCC_STAND_ALONE
-    // TODO use sys.argv ?
+#ifdef Q_OS_WIN32
+    const py::module_ os = py::module_::import("os");
+    const py::module_ sysconfig = py::module_::import("sysconfig");
+    const QString site_packages = sysconfig.attr("get_path")("purelib").cast<QString>();
+    const QString platform_path = QString("%1\\pycc\\platforms").arg(site_packages);
+    const std::string std_platform_path = platform_path.toStdString();
+    os.attr("environ")["QT_QPA_PLATFORM_PLUGIN_PATH"] = std_platform_path;
+#endif
+    // TODO use sys.argv ? and if then we don't need the env var thing on windows
     static char fakeArg0[] = "pycc";
     static char *fakeArgv[] = {&fakeArg0[0], NULL};
     static int argc = 1;
@@ -95,61 +91,6 @@ void InitializeQApplication()
     // so, we have nothing to do.
 }
 
-void load_pluginWrappers(py::module_ &m)
-{
-    py::module_ pluginWrappers = m.def_submodule("plugins");
-    pluginWrappers.doc() = "Module we wrappers around some CloudCompare plugins";
-
-#ifdef PYCC_STAND_ALONE
-    py::module_ sys = py::module_::import("sys");
-    const QString envPrefix = sys.attr("prefix").cast<QString>();
-    const QString platlibdir = guessPlatlibdir();
-    QString pluginPath = QString("%1/%2/cloudcompare/plugins-python").arg(envPrefix, platlibdir);
-#else
-    // TODO it might be interesting to get the lib64 / lib from a config.h generated by cmake
-    //  instead of calling python
-    const QString platlibdir = guessPlatlibdir();
-    QDir theDir = QCoreApplication::applicationDirPath();
-    if (theDir.dirName() == "bin")
-    {
-        theDir.cdUp();
-    }
-
-    QString pluginPath = QString("%1/%2/cloudcompare/plugins-python").arg(theDir.absolutePath(), platlibdir);
-#endif
-
-    QDir pluginPathDir(pluginPath);
-    if (!pluginPathDir.exists())
-    {
-        return;
-    }
-
-    py::module_::import("sys").attr("path").attr("append")(pluginPath);
-    QDirIterator iter(pluginPathDir);
-    while (iter.hasNext())
-    {
-        iter.next();
-        QFileInfo entry = iter.fileInfo();
-        QString fileName = entry.fileName();
-
-        if (!fileName.endsWith(".so"))
-        {
-            continue;
-        }
-
-        // Our wrappers are wheels with name like "pluginName.cpython.etc.etc.so"
-        const std::string stdFileName = fileName.split('.')[0].toStdString();
-        try {
-            const py::module mod = py::module_::import(stdFileName.c_str());
-
-            const std::string pluginName = mod.attr("plugin_name").cast<std::string>();
-            pluginWrappers.attr(pluginName.c_str()) = mod;
-        } catch (const std::exception& e) {
-            py::print("Failed to load plugin '", stdFileName, "': ", e.what());
-            continue;
-        }
-    }
-}
 
 void import_pyccRuntime(py::module &m)
 {
